+++ docs/src/argo_trajectory.jl
@@ -1,5 +1,5 @@
 # Plot a float trajectory with colour for sequence number
-using OceanAnalysis, Plots, Statistics
+using OceanAnalysis, Plots, Printf, Statistics
 ID = r"D4902911" # focus on this ID
 index_file = get_argo_index("~/data/argo");
 index_all = read_argo_index(index_file) # 3.2e6 profiles
@@ -9,7 +9,8 @@ lon, lat = index.longitude, index.latitude
 plot(lon, lat,
     aspect_ratio=1.0 / cos(mean(lat) * pi / 180),
     framestyle=:box, color=:gray, dpi=200,
-    title="Argo float $(ID.pattern) coloured by cycle index", titlefontsize=9)
+    title=@sprintf("Argo float %s coloured by cycle index", ID.pattern),
+    titlefontsize=9)
 colors = cgrad(:turbo)
 scatter!(lon, lat, marker_z=1:length(lon),
     markersize=3, markerstyle=:circle, color=colors)